Dashboard/requirements.txt
asgiref==3.3.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
distro==1.4.0
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
keyring==18.0.1
keyrings.alt==3.4.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyfcm==1.4.9
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
systemd-python==234
urllib3==1.25.8
wadllib==1.3.3